% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.celltype.data.R
\name{generate.celltype.data}
\alias{generate.celltype.data}
\title{generate.celltype.data}
\usage{
generate.celltype.data(
  exp,
  annot,
  GroupName,
  run_sct_transform = TRUE,
  numCores = 1
)
}
\arguments{
\item{exp}{Numerical matrix with row for each gene and column for each cell. Row names are MGI/HGNC gene symbols. Column names are cell IDs which can be cross referenced against the annot data frame.}

\item{annot}{dataframe with columns containing the cell type names associated with each column in exp. Requires "patient_id" and "pathology" columns.}

\item{numCores}{Number of cores that should be used to speedup the computation}

\item{groupName}{A human readable name for refering to the dataset being loaded}
}
\value{
Filenames for the saved celltype_data files
}
\description{
\code{generate.celltype.data} Takes expression & cell type annotations and creates celltype_data files which contain the mean and specificity matrices
}
\examples{
# Load the single cell data
data("cortex_mrna")
expData = cortex_mrna$exp
fNames_ALLCELLS = generate.celltype.data(exp=expData,annotLevels,"allKImouse")
}
